fix: 간헐적으로 refresh가 이루어지지 않는 버그
