+++ libs/client/AuthProvider.tsx
@@ -1,16 +1,12 @@
 'use client';
 
 import { useCallback, useEffect } from 'react';
-import { InternalAxiosRequestConfig } from 'axios';
-import client from '@/libs/client/client';
 import { useGetProfile } from '@/libs/apis/users';
 import { useGetTickets } from '../apis/tickets';
 import { useAuthStore } from '../store/Providers/AuthStoreProvider';
 import { useProfileStore } from '../store/Providers/ProfileStoreProvider';
 import { useTicketStore } from '../store/Providers/TicketStoreProvider';
-import mem from 'mem';
-
-const REFRESH_URL = '/auth/refresh';
+import { refresh } from '@/libs/client/createAxiosInstance';
 
 export function AuthProvider() {
   const { isLogin, login, logout } = useAuthStore((state) => state);
@@ -33,72 +29,6 @@ export function AuthProvider() {
     localStorage.removeItem('refreshToken');
     logout();
   }, [logout]);
-
-  const refresh = useCallback(
-    mem(
-      async () => {
-        try {
-          const { data } = await client.post(REFRESH_URL);
-          if (data.accessToken && data.refreshToken) {
-            signIn(data.accessToken, data.refreshToken);
-            return data.accessToken as string;
-          }
-        } catch (err) {
-          console.log(err);
-          signOut();
-        }
-
-        return null;
-      },
-      { maxAge: 1000 },
-    ),
-    [signIn, signOut],
-  );
-
-  const errorHandler = async (error: any) => {
-    const { config, response } = error;
-
-    if (response?.status === 401 && config.url !== REFRESH_URL && !config.sent) {
-      config.sent = true;
-      const newAccessToken = await refresh();
-
-      if (newAccessToken) {
-        console.log(`New Access Token :: ${newAccessToken}`);
-        config.headers.Authorization = `Bearer ${newAccessToken}`;
-        return client(config);
-      } else {
-        return Promise.reject(error);
-      }
-    }
-
-    return Promise.reject(error);
-  };
-
-  const requestHandler = (config: InternalAxiosRequestConfig) => {
-    let token: string | null = null;
-
-    if (config.url === REFRESH_URL) {
-      token = localStorage.getItem('refreshToken');
-    } else {
-      token = localStorage.getItem('accessToken');
-    }
-
-    if (token !== null) {
-      config.headers.Authorization = `Bearer ${token}`;
-    }
-
-    return config;
-  };
-
-  const requestInterceptor = client.interceptors.request.use(requestHandler, (error) => Promise.reject(error));
-  const responseInterceptor = client.interceptors.response.use((response) => response, errorHandler);
-
-  useEffect(() => {
-    return () => {
-      client.interceptors.request.eject(requestInterceptor);
-      client.interceptors.response.eject(responseInterceptor);
-    };
-  }, [responseInterceptor, requestInterceptor]);
 
   useEffect(() => {
     const accessToken = localStorage.getItem('accessToken');
@@ -125,7 +55,7 @@ export function AuthProvider() {
         refresh();
       }
     }
-  }, [signIn, signOut, refresh]);
+  }, [signIn, signOut]);
 
   useEffect(() => {
     const refreshToken = localStorage.getItem('refreshToken');
